supply_chain_predictor/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data processing
pandas==2.1.4
numpy==1.26.3

# Time series forecasting
prophet==1.1.5

# Data validation
pydantic==2.5.3

# LLM integration
groq==0.4.1

# Environment configuration
python-dotenv==1.0.0

# Additional dependencies for Prophet
cmdstanpy==1.2.0
holidays==0.40

# Utility packages
httpx==0.26.0

